refactor: wip set option through toggle
